Extracting puzzle game component
